pyproject.toml
[project]
name = "credit-decisioning-platform"
version = "0.1.0"
description = "Decision Intelligence credit decisioning platform"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.2",
  "numpy>=1.26",
  "scikit-learn>=1.4",
  "lightgbm>=4.3",
  "pydantic>=2.6",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "joblib>=1.3",
  "python-dotenv>=1.0",
  "requests>=2.32",
  "pyarrow>=16.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "black>=24.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
